refactor: preparing changes for voice querying
